+++ scipy/misc/__init__.py
@@ -16,7 +16,7 @@ systems that don't have PIL installed.
 
    bytescale - Byte scales an array (image)
    central_diff_weights - Weights for an n-point central m-th derivative
-   comb - Combinations of N things taken k at a time, "N choose k"
+   comb - Combinations of N things taken k at a time, "N choose k" (imported from scipy.special)
    derivative - Find the n-th derivative of a function at a point
    factorial  - The factorial function, n! = special.gamma(n+1)
    factorial2 - Double factorial, (n!)!
@@ -32,7 +32,7 @@ systems that don't have PIL installed.
    lena - Get classic image processing example image Lena
    logsumexp - Compute the log of the sum of exponentials of input elements
    pade - Pade approximation to function as the ratio of two polynomials
-   perm - Permutations of N things taken k at a time, "k-permutations of N"
+   perm - Permutations of N things taken k at a time, "k-permutations of N" (imported from scipy.special)
    toimage - Takes a numpy array and returns a PIL image
    who - Print the Numpy arrays in the given dictionary
 
@@ -45,6 +45,7 @@ __all__ = ['who', 'source', 'info', 'doccer']
 from . import doccer
 from .common import *
 from numpy import who, source, info as _info
+from scipy.special import perm, comb
 
 import sys
